requirements.txt
dwave_neal==0.6.0
numpy==2.1.2
scikit_learn==1.4.2
